Fixes Ruby 2.7 warnings.
I don't believe that you can override the constructor of a Struct, though it is weird that it doesn't break things. The code becomes torn between wanting to accept keyword arguments and expecting ordered arguments and ultimately complains about it through the warning.

Turning the class from a Struct into a regular class means we can use keyword arguments in the constructor with no problem.